Spawn a single terminal
This commit implements the base functionality of the termmaker plugin.
It allows to create terminals and toggle them on or off. For the moment
the window from which the terminal was opened is re-used to display the
terminal buffer. The original values of any modified window options are
saved. The window options are restored upon toggling the terminal off or
loading another buffer into the window displaying the terminal.